requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.12.13
aiosignal==1.3.2
annotated-types==0.7.0
antlr4-python3-runtime==4.9.3
attrs==25.3.0
certifi==2025.4.26
charset-normalizer==3.4.2
click==8.2.1
contourpy==1.3.2
cycler==0.12.1
decorator==5.2.1
diffusers==0.33.1
dill==0.4.0
einops==0.8.1
ema-pytorch==0.7.7
filelock==3.18.0
fonttools==4.58.4
frozenlist==1.7.0
fsspec==2025.5.1
gitdb==4.0.12
GitPython==3.1.44
hf-xet==1.1.3
huggingface-hub==0.33.0
hydra-core==1.3.2
idna==3.10
importlib_metadata==8.7.0
Jinja2==3.1.6
joblib==1.5.1
kiwisolver==1.4.8
lazy_loader==0.4
lightning==2.5.1.post0
lightning-utilities==0.14.3
MarkupSafe==3.0.2
matplotlib==3.10.3
mne==1.9.0
mpmath==1.3.0
multidict==6.4.4
networkx==3.5
numpy==1.26.4
nvidia-cublas-cu12==12.6.4.1
nvidia-cuda-cupti-cu12==12.6.80
nvidia-cuda-nvrtc-cu12==12.6.77
nvidia-cuda-runtime-cu12==12.6.77
nvidia-cudnn-cu12==9.5.1.17
nvidia-cufft-cu12==11.3.0.4
nvidia-cufile-cu12==1.11.1.6
nvidia-curand-cu12==10.3.7.77
nvidia-cusolver-cu12==11.7.1.2
nvidia-cusparse-cu12==12.5.4.2
nvidia-cusparselt-cu12==0.6.3
nvidia-nccl-cu12==2.26.2
nvidia-nvjitlink-cu12==12.6.85
nvidia-nvtx-cu12==12.6.77
omegaconf==2.3.0
packaging==24.2
pandarallel==1.6.5
pandas==1.5.3
pillow==11.2.1
pip==25.1
platformdirs==4.3.8
pooch==1.8.2
propcache==0.3.2
protobuf==6.31.1
psutil==7.0.0
pydantic==2.11.7
pydantic_core==2.33.2
pyhealth==1.1.6
pyparsing==3.2.3
python-dateutil==2.9.0.post0
pytorch-lightning==2.5.1.post0
pytz==2025.2
PyYAML==6.0.2
rdkit==2025.3.3
regex==2024.11.6
requests==2.32.4
safetensors==0.5.3
scikit-learn==1.7.0
scipy==1.15.3
sentry-sdk==2.30.0
setproctitle==1.3.6
setuptools==78.1.1
six==1.17.0
smmap==5.0.2
sympy==1.14.0
threadpoolctl==3.6.0
torch==2.7.1
torchaudio==2.7.1
torchmetrics==1.7.3
torchvision==0.22.1
tqdm==4.67.1
triton==3.3.1
typing_extensions==4.14.0
typing-inspection==0.4.1
urllib3==1.26.15
wandb==0.20.1
wheel==0.45.1
yarl==1.20.1
zipp==3.23.0
